manage errors & exceptions
