+++ handlers/vpc_handler.go
@@ -1,0 +1,127 @@
+package handlers
+
+import (
+	"net/http"
+
+	cache "github.com/patrickmn/go-cache"
+)
+
+func (handler *AWSHandler) VPCTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("vpc_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeVPCsTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("vpc_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) ACLTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("acl_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeACLsTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("acl_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) SecurityGroupTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("sg_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeSecurityGroupsTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("sg_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) NatGatewayTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("nat_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeNatGatewaysTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("nat_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) ElasticIPTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("eip_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeElasticIPsTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("eip_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) InternetGatewayTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("igw_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeInternetGatewaysTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("igw_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) RouteTableTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("rt_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeRouteTablesTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("rt_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
+
+func (handler *AWSHandler) KeyPairTotalHandler(w http.ResponseWriter, r *http.Request) {
+	response, found := handler.cache.Get("kp_total")
+	if found {
+		respondWithJSON(w, 200, response)
+	} else {
+		response, err := handler.aws.DescribeKeyPairsTotal(handler.cfg)
+		if err != nil {
+			respondWithError(w, http.StatusInternalServerError, "You dont have the right permission")
+		} else {
+			handler.cache.Set("kp_total", response, cache.DefaultExpiration)
+			respondWithJSON(w, 200, response)
+		}
+	}
+}
